refactor(styles): split interactive.styles.ts (pass 1)
Extract the two layers that share nothing with Emotion CSS into their
own files; keep interactive.styles.ts as a thin barrel so every consumer
import continues to resolve unchanged. Pure mechanical move — verified
via byte-exact diff between the original line ranges and the relocated
content; no logic, formatting, selector, or CSS variable changes.

  - interactive.theme-bridge.ts  (88 lines)  updateInteractiveThemeColors
  - interactive.overlay.styles.ts (944 lines) addGlobalInteractiveStyles
  - interactive.styles.ts        (1209 lines) helpers + getInteractiveStyles + re-exports
