+++ src/styles/interactive.styles.ts
@@ -1,94 +1,8 @@
 import { css } from '@emotion/css';
 import { GrafanaTheme2 } from '@grafana/data';
-import { INTERACTIVE_CONFIG } from '../constants/interactive-config';
-import { INTERACTIVE_Z_INDEX } from '../constants/interactive-z-index';
 
-/**
- * Updates CSS custom properties for interactive comment boxes based on the current theme.
- * This function should be called whenever the theme might change to ensure the comment
- * box popups (which are rendered outside the React tree as global DOM elements) respect
- * the user's light/dark mode preference.
- *
- * The custom properties are written to :root (document.documentElement) and intentionally
- * persist — no cleanup is required. They are namespaced under --pathfinder-* to avoid
- * collisions with Grafana's own CSS variables.
- *
- * @param theme - The current Grafana theme
- */
-export const updateInteractiveThemeColors = (theme: GrafanaTheme2): void => {
-  const root = document.documentElement;
-  const isDark = theme.isDark;
-
-  // Background and text colors
-  root.style.setProperty('--pathfinder-comment-bg', theme.colors.background.primary);
-  root.style.setProperty('--pathfinder-comment-text', theme.colors.text.primary);
-  root.style.setProperty('--pathfinder-comment-text-secondary', theme.colors.text.secondary);
-  root.style.setProperty('--pathfinder-comment-border-weak', theme.colors.border.weak);
-
-  // UI element colors (buttons, close icons, etc.)
-  // These need different opacities for light vs dark mode to maintain contrast
-  root.style.setProperty(
-    '--pathfinder-comment-close-color',
-    isDark ? 'rgba(255, 255, 255, 0.5)' : 'rgba(0, 0, 0, 0.4)'
-  );
-  root.style.setProperty(
-    '--pathfinder-comment-close-hover-bg',
-    isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.08)'
-  );
-  root.style.setProperty(
-    '--pathfinder-comment-close-hover-color',
-    isDark ? 'rgba(255, 255, 255, 0.9)' : 'rgba(0, 0, 0, 0.8)'
-  );
-
-  // Button and input borders
-  root.style.setProperty(
-    '--pathfinder-comment-btn-border',
-    isDark ? 'rgba(255, 255, 255, 0.2)' : 'rgba(0, 0, 0, 0.15)'
-  );
-  root.style.setProperty('--pathfinder-comment-btn-text', isDark ? 'rgba(255, 255, 255, 0.8)' : 'rgba(0, 0, 0, 0.7)');
-  root.style.setProperty(
-    '--pathfinder-comment-btn-hover-bg',
-    isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.06)'
-  );
-  root.style.setProperty(
-    '--pathfinder-comment-btn-hover-border',
-    isDark ? 'rgba(255, 255, 255, 0.3)' : 'rgba(0, 0, 0, 0.25)'
-  );
-  root.style.setProperty('--pathfinder-comment-btn-hover-text', isDark ? '#ffffff' : '#000000');
-
-  // Progress and indicator colors
-  root.style.setProperty(
-    '--pathfinder-comment-progress-bg',
-    isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.08)'
-  );
-  root.style.setProperty('--pathfinder-comment-dot-bg', isDark ? 'rgba(255, 255, 255, 0.2)' : 'rgba(0, 0, 0, 0.15)');
-
-  // Code element colors
-  root.style.setProperty('--pathfinder-comment-code-bg', isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.06)');
-  root.style.setProperty(
-    '--pathfinder-comment-code-border',
-    isDark ? 'rgba(255, 255, 255, 0.2)' : 'rgba(0, 0, 0, 0.12)'
-  );
-
-  // Keyboard hints
-  root.style.setProperty('--pathfinder-comment-kbd-text', isDark ? 'rgba(255, 255, 255, 0.35)' : 'rgba(0, 0, 0, 0.35)');
-  root.style.setProperty('--pathfinder-comment-kbd-bg', isDark ? 'rgba(255, 255, 255, 0.08)' : 'rgba(0, 0, 0, 0.05)');
-  root.style.setProperty(
-    '--pathfinder-comment-kbd-border',
-    isDark ? 'rgba(255, 255, 255, 0.12)' : 'rgba(0, 0, 0, 0.1)'
-  );
-
-  // Border top for button container
-  root.style.setProperty('--pathfinder-comment-divider', isDark ? 'rgba(255, 255, 255, 0.08)' : 'rgba(0, 0, 0, 0.08)');
-
-  // Form validation status colors
-  root.style.setProperty('--pathfinder-form-checking-bg', isDark ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.05)');
-  root.style.setProperty(
-    '--pathfinder-form-checking-border',
-    isDark ? 'rgba(255, 255, 255, 0.2)' : 'rgba(0, 0, 0, 0.12)'
-  );
-  root.style.setProperty('--pathfinder-form-checking-text', isDark ? 'rgba(255, 255, 255, 0.8)' : 'rgba(0, 0, 0, 0.7)');
-};
+export { updateInteractiveThemeColors } from './interactive.theme-bridge';
+export { addGlobalInteractiveStyles } from './interactive.overlay.styles';
 
 // Base interactive element styles
 const getBaseInteractiveStyles = (theme: GrafanaTheme2) => ({
@@ -1293,945 +1207,3 @@ export const getInteractiveStyles = (theme: GrafanaTheme2) =>
     ...getInteractiveComponentStyles(theme),
     ...getExpandableStyles(theme),
   } as any);
-
-// Pure global (vanilla) CSS for overlays/highlights—run once at app startup
-export const addGlobalInteractiveStyles = () => {
-  const interactiveStyleId = 'interactive-global-styles';
-  if (document.getElementById(interactiveStyleId)) {
-    return;
-  }
-  const style = document.createElement('style');
-  style.id = interactiveStyleId;
-
-  // ============================================================
-  // ANIMATION TIMING CONSTANTS
-  // ============================================================
-  // These must stay synchronized with INTERACTIVE_CONFIG.highlighting in interactive-config.ts
-  // See that file for the authoritative timing documentation.
-  //
-  // DOT INDICATOR:
-  //   CSS: breatheCycleMs × breatheCount (3.0s) + fadeMs (0.5s) = 3.5s total
-  //   JS:  dotDurationMs = 4000ms (500ms buffer for cleanup)
-  //
-  // BOUNDING BOX OUTLINE:
-  //   drawMs = Math.max(500, Math.round(highlightMs * 0.65)) = 1625ms (when highlightMs = 2500)
-  //   CSS: drawMs (1625ms) + breatheCycleMs × breatheCount (3.0s) + fadeMs (0.5s) = 5.125s total
-  //   JS:  outlineDurationMs = 5625ms (500ms buffer for cleanup)
-  // ============================================================
-  const breatheCycleMs = 1500;
-  const breatheCount = 2;
-  const fadeMs = 500;
-  const breatheTotalMs = breatheCycleMs * breatheCount; // 3000ms
-
-  // Align highlight animation timing with configured technical highlight delay
-  const highlightMs = INTERACTIVE_CONFIG.delays.technical.highlight;
-  // Slower, more readable draw animation (no fade-out since highlights persist)
-  const drawMs = Math.max(500, Math.round(highlightMs * 0.65));
-  style.textContent = `
-    /* Blocker overlay visuals (used by GlobalInteractionBlocker) */
-    #interactive-blocking-overlay {
-      background: transparent !important;
-      /* Always-visible subtle border */
-      border: 1px solid rgba(170, 170, 170, 0.35);
-      /* Breathing pulse layered on top */
-      box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.22);
-      animation: blocker-breathe 3.2s ease-in-out infinite;
-    }
-
-    /* When a modal is active, remove breathing pulse to avoid visual clash and keep overlay subtle */
-    #interactive-blocking-overlay.no-breathe {
-      animation: none !important;
-      box-shadow: none !important;
-      border: none !important;
-    }
-
-    /* Full-screen overlay for modal blocking */
-    #interactive-fullscreen-overlay {
-      background: transparent !important;
-      /* Gray pulse around the outside edge - more visible for full screen */
-      border: 2px solid rgba(170, 170, 170, 0.4);
-      box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.3);
-      animation: fullscreen-breathe 2.8s ease-in-out infinite;
-    }
-
-    /* Header overlay styling */
-    #interactive-header-overlay {
-      background: transparent !important;
-      border: 1px solid rgba(170, 170, 170, 0.35);
-      box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.22);
-      animation: blocker-breathe 3.2s ease-in-out infinite;
-    }
-
-    @keyframes blocker-breathe {
-      0% {
-        box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.18);
-      }
-      50% {
-        box-shadow: inset 0 0 0 6px rgba(170, 170, 170, 0.20);
-      }
-      100% {
-        box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.18);
-      }
-    }
-
-    @keyframes fullscreen-breathe {
-      0% {
-        box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.25);
-        border-color: rgba(170, 170, 170, 0.35);
-      }
-      50% {
-        box-shadow: inset 0 0 0 8px rgba(170, 170, 170, 0.30);
-        border-color: rgba(170, 170, 170, 0.45);
-      }
-      100% {
-        box-shadow: inset 0 0 0 0 rgba(170, 170, 170, 0.25);
-        border-color: rgba(170, 170, 170, 0.35);
-      }
-    }
-    /* Global interactive highlight styles */
-    .interactive-highlight-outline {
-      position: fixed;
-      top: var(--highlight-top);
-      left: var(--highlight-left);
-      width: var(--highlight-width);
-      height: var(--highlight-height);
-      pointer-events: none;
-      z-index: ${INTERACTIVE_Z_INDEX.HIGHLIGHT_OUTLINE};
-      border-radius: 4px;
-      /* Draw border clockwise using four gradient strokes (no fill) */
-      --hl-color: rgba(255, 136, 0, 0.85);
-      --hl-thickness: 2px;
-      background:
-        linear-gradient(var(--hl-color) 0 0) top left / 0 var(--hl-thickness) no-repeat,
-        linear-gradient(var(--hl-color) 0 0) top right / var(--hl-thickness) 0 no-repeat,
-        linear-gradient(var(--hl-color) 0 0) bottom right / 0 var(--hl-thickness) no-repeat,
-        linear-gradient(var(--hl-color) 0 0) bottom left / var(--hl-thickness) 0 no-repeat;
-      opacity: 0.95;
-      /* Draw border animation, then breathing glow (2 cycles), then fade out
-       * Total: 1625ms draw + 3s breathe + 0.5s fade = 5.125s
-       * JS cleanup at 5625ms (500ms buffer after CSS completes)
-       */
-      animation:
-        interactive-draw-border ${drawMs}ms cubic-bezier(0.18, 0.6, 0.2, 1) forwards,
-        interactive-glow-breathe ${breatheCycleMs}ms ease-in-out ${drawMs}ms ${breatheCount},
-        interactive-outline-fade ${fadeMs}ms ease-out ${drawMs + breatheTotalMs}ms forwards;
-    }
-
-    /* Subtle variant to reuse animation cadence for blocked areas */
-    .interactive-highlight-outline--subtle {
-      border-color: rgba(180, 180, 180, 0.4);
-      background-color: rgba(180, 180, 180, 0.08);
-      box-shadow: 0 0 0 4px rgba(180, 180, 180, 0.12);
-      animation: subtle-highlight-pulse 1.6s ease-in-out infinite;
-    }
-
-    /* Hover highlight for element inspector (watch/record mode) */
-    .interactive-hover-highlight-outline {
-      position: fixed;
-      top: var(--highlight-top);
-      left: var(--highlight-left);
-      width: var(--highlight-width);
-      height: var(--highlight-height);
-      pointer-events: none;
-      z-index: ${INTERACTIVE_Z_INDEX.HIGHLIGHT_OUTLINE};
-      border-radius: 4px;
-      /* Same orange color as regular highlights for consistency */
-      --hl-color: rgba(255, 136, 0, 0.85);
-      --hl-thickness: 2px;
-      background:
-        linear-gradient(var(--hl-color) 0 0) top left / 0 var(--hl-thickness) no-repeat,
-        linear-gradient(var(--hl-color) 0 0) top right / var(--hl-thickness) 0 no-repeat,
-        linear-gradient(var(--hl-color) 0 0) bottom right / 0 var(--hl-thickness) no-repeat,
-        linear-gradient(var(--hl-color) 0 0) bottom left / var(--hl-thickness) 0 no-repeat;
-      /* No animation for hover - instant feedback */
-      background-size: 100% var(--hl-thickness), var(--hl-thickness) 100%, 100% var(--hl-thickness), var(--hl-thickness) 100%;
-      opacity: 0.95;
-      /* Smooth transitions when moving between elements */
-      transition: top 0.05s ease-out, left 0.05s ease-out, width 0.05s ease-out, height 0.05s ease-out;
-    }
-
-    @keyframes interactive-draw-border {
-      0% {
-        background-size: 0 var(--hl-thickness), var(--hl-thickness) 0, 0 var(--hl-thickness), var(--hl-thickness) 0;
-      }
-      25% {
-        background-size: 100% var(--hl-thickness), var(--hl-thickness) 0, 0 var(--hl-thickness), var(--hl-thickness) 0;
-      }
-      50% {
-        background-size: 100% var(--hl-thickness), var(--hl-thickness) 100%, 0 var(--hl-thickness), var(--hl-thickness) 0;
-      }
-      75% {
-        background-size: 100% var(--hl-thickness), var(--hl-thickness) 100%, 100% var(--hl-thickness), var(--hl-thickness) 0;
-      }
-      100% {
-        background-size: 100% var(--hl-thickness), var(--hl-thickness) 100%, 100% var(--hl-thickness), var(--hl-thickness) 100%;
-      }
-    }
-
-    /* Breathing orange glow that activates after border draw completes */
-    @keyframes interactive-glow-breathe {
-      0%, 100% {
-        box-shadow: 0 0 8px 2px rgba(255, 136, 0, 0.3);
-      }
-      50% {
-        box-shadow: 0 0 16px 4px rgba(255, 136, 0, 0.5);
-      }
-    }
-
-    /* Fade out animation for bounding box after breathing completes */
-    @keyframes interactive-outline-fade {
-      from {
-        opacity: 0.95;
-        box-shadow: 0 0 8px 2px rgba(255, 136, 0, 0.3);
-      }
-      to {
-        opacity: 0;
-        box-shadow: 0 0 4px 1px rgba(255, 136, 0, 0);
-      }
-    }
-
-    /* Instant highlight - no draw animation, just breathing glow then fade
-     * Instant: 2 breathe cycles (3s) + 0.5s fade = 3.5s
-     */
-    .interactive-highlight-outline--instant {
-      animation: 
-        interactive-glow-breathe ${breatheCycleMs}ms ease-in-out ${breatheCount},
-        interactive-outline-fade ${fadeMs}ms ease-out ${breatheTotalMs}ms forwards !important;
-      background-size: 100% var(--hl-thickness), var(--hl-thickness) 100%, 100% var(--hl-thickness), var(--hl-thickness) 100% !important;
-    }
-
-    /* Pulsing dot indicator for small/hidden elements (< 10px width or height)
-     * Threshold defined in INTERACTIVE_CONFIG.highlighting.minDimensionForBox
-     */
-    .interactive-highlight-dot {
-      position: fixed;
-      top: var(--highlight-top);
-      left: var(--highlight-left);
-      width: 16px;
-      height: 16px;
-      border-radius: 50%;
-      background: rgba(255, 136, 0, 0.9);
-      pointer-events: none;
-      z-index: ${INTERACTIVE_Z_INDEX.HIGHLIGHT_OUTLINE};
-      transform: translate(-50%, -50%); /* Center on the point */
-      box-shadow: 0 0 12px 4px rgba(255, 136, 0, 0.5);
-      /* Finite animation: 2 pulse cycles (3s) then fade out (0.5s) = 3.5s total
-       * JS cleanup at 4000ms (500ms buffer after CSS completes)
-       */
-      animation: 
-        interactive-dot-pulse ${breatheCycleMs}ms ease-in-out ${breatheCount},
-        interactive-dot-fade ${fadeMs}ms ease-out ${breatheTotalMs}ms forwards;
-    }
-
-    @keyframes interactive-dot-pulse {
-      0%, 100% {
-        transform: translate(-50%, -50%) scale(1);
-        opacity: 0.9;
-      }
-      50% {
-        transform: translate(-50%, -50%) scale(1.3);
-        opacity: 1;
-      }
-    }
-
-    @keyframes interactive-dot-fade {
-      from {
-        opacity: 0.9;
-        transform: translate(-50%, -50%) scale(1);
-      }
-      to {
-        opacity: 0;
-        transform: translate(-50%, -50%) scale(0.8);
-      }
-    }
-
-    /* Enhanced comment box animations */
-    @keyframes fadeInComment {
-      0% {
-        opacity: 0;
-        transform: scale(0.85) translateY(-8px);
-      }
-      60% {
-        transform: scale(1.02) translateY(0);
-      }
-      100% {
-        opacity: 1;
-        transform: scale(1) translateY(0);
-      }
-    }
-
-    /* Comment box exit animation */
-    .comment-box-exit {
-      animation: fadeOutComment 0.2s ease-in forwards !important;
-    }
-
-    @keyframes fadeOutComment {
-      0% {
-        opacity: 1;
-        transform: scale(1) translateY(0);
-      }
-      100% {
-        opacity: 0;
-        transform: scale(0.9) translateY(-5px);
-      }
-    }
-
-    /* Spinner animation for loading states */
-    @keyframes spin {
-      to {
-        transform: rotate(360deg);
-      }
-    }
-
-    /* Progress bar pulse animation */
-    @keyframes progressPulse {
-      0%, 100% {
-        opacity: 1;
-      }
-      50% {
-        opacity: 0.6;
-      }
-    }
-
-    @keyframes subtle-highlight-pulse {
-      0% {
-        opacity: 0.55;
-        transform: scale(0.995);
-        box-shadow: 0 0 0 0 rgba(180, 180, 180, 0.12);
-      }
-      50% {
-        opacity: 0.8;
-        transform: scale(1);
-        box-shadow: 0 0 0 6px rgba(180, 180, 180, 0.16);
-      }
-      100% {
-        opacity: 0.55;
-        transform: scale(0.995);
-        box-shadow: 0 0 0 0 rgba(180, 180, 180, 0.12);
-      }
-    }
-
-    /* Fragment highlighting for anchor navigation */
-    .fragment-highlight {
-      position: relative;
-      background-color: rgba(255, 193, 7, 0.2) !important;
-      border-left: 4px solid #ffc107 !important;
-      padding-left: 8px !important;
-      margin-left: -12px !important;
-      animation: fragment-highlight-fade 3s ease-out forwards;
-    }
-
-    @keyframes fragment-highlight-fade {
-      0% {
-        background-color: rgba(255, 193, 7, 0.4);
-        border-left-color: #ffc107;
-      }
-      50% {
-        background-color: rgba(255, 193, 7, 0.3);
-        border-left-color: #ffc107;
-      }
-      100% {
-        background-color: rgba(255, 193, 7, 0.1);
-        border-left-color: transparent;
-      }
-    }
-
-    /* Interactive comment box - body-attached, absolute positioning
-     * top and left are set directly in JS via inline styles
-     */
-    .interactive-comment-box {
-      position: fixed;
-      /* top and left are set via inline styles in navigation-manager.ts */
-      width: 420px;
-      max-width: calc(100vw - 32px);
-      pointer-events: none;
-      z-index: ${INTERACTIVE_Z_INDEX.COMMENT_BOX};
-      /* Initial state - hidden with slight offset for slide-in effect */
-      opacity: 0;
-      transform: scale(0.96);
-      transition: opacity 0.25s cubic-bezier(0.4, 0, 0.2, 1),
-                  transform 0.25s cubic-bezier(0.4, 0, 0.2, 1);
-    }
-
-    /* Slide-in from different directions based on position */
-    .interactive-comment-box[data-position="right"] {
-      transform: scale(0.96) translateX(-8px);
-    }
-    .interactive-comment-box[data-position="left"] {
-      transform: scale(0.96) translateX(8px);
-    }
-    .interactive-comment-box[data-position="bottom"] {
-      transform: scale(0.96) translateY(-8px);
-    }
-    .interactive-comment-box[data-position="top"] {
-      transform: scale(0.96) translateY(8px);
-    }
-
-    /* Centered position for noop steps (no element to highlight) */
-    .interactive-comment-box[data-position="center"] {
-      position: fixed;
-      top: 50%;
-      left: 50%;
-      transform: scale(0.96) translate(-50%, -50%);
-      z-index: 10001;
-    }
-
-    /* Noop comment boxes get a slightly different style */
-    .interactive-comment-box[data-noop="true"] .interactive-comment-content::before {
-      display: none; /* No arrow for centered boxes */
-    }
-
-    /* Final state - visible and in position */
-    /* Override position-specific transforms to prevent overlap */
-    .interactive-comment-box[data-ready="true"][data-position="right"],
-    .interactive-comment-box[data-ready="true"][data-position="left"],
-    .interactive-comment-box[data-ready="true"][data-position="top"],
-    .interactive-comment-box[data-ready="true"][data-position="bottom"] {
-      opacity: 1;
-      transform: scale(1) translateX(0) translateY(0);
-    }
-
-    .interactive-comment-box[data-ready="true"][data-position="center"] {
-      transform: scale(1) translate(-50%, -50%);
-    }
-
-    /* Instant comment box - no transition, final state immediately */
-    /* Use high specificity to override position-specific transforms */
-    .interactive-comment-box.interactive-comment-box--instant,
-    .interactive-comment-box.interactive-comment-box--instant[data-position="right"],
-    .interactive-comment-box.interactive-comment-box--instant[data-position="left"],
-    .interactive-comment-box.interactive-comment-box--instant[data-position="top"],
-    .interactive-comment-box.interactive-comment-box--instant[data-position="bottom"] {
-      transition: none !important;
-      opacity: 1 !important;
-      transform: none !important;
-    }
-
-    .interactive-comment-box.interactive-comment-box--instant[data-position="center"] {
-      transform: translate(-50%, -50%) !important;
-    }
-
-    .interactive-comment-box--instant .interactive-comment-content {
-      animation: none !important;
-      box-shadow: 
-        0 4px 20px rgba(0, 0, 0, 0.25),
-        0 0 0 2px rgba(255, 136, 0, 0.25),
-        0 0 15px rgba(255, 136, 0, 0.15) !important;
-    }
-
-    .interactive-comment-content {
-      border-radius: 8px;
-      padding: 12px;
-      padding-right: 36px; /* Extra padding for close button */
-      font-size: 14px;
-      line-height: 1.5;
-      /* Layered shadow: soft glow + depth shadow */
-      box-shadow: 
-        0 0 20px rgba(255, 152, 0, 0.15),
-        0 4px 20px rgba(0, 0, 0, 0.25),
-        0 1px 3px rgba(0, 0, 0, 0.1);
-      position: relative;
-      /* Essential styling that uses CSS custom properties set by updateInteractiveThemeColors().
-         Fallback values (#1f1f23, #d9d9d9) are dark-mode defaults — they apply only in the
-         brief window before the first React effect fires. updateInteractiveThemeColors() must
-         have run before a popup is shown to guarantee correct colors in light mode. */
-      background: var(--pathfinder-comment-bg, #1f1f23);
-      border: 1px solid rgba(255, 152, 0, 0.3);
-      color: var(--pathfinder-comment-text, #d9d9d9);
-      /* Ensure content fits within container bounds */
-      overflow: hidden;
-      word-wrap: break-word;
-      overflow-wrap: break-word;
-      max-width: 100%;
-      box-sizing: border-box;
-    }
-
-    /* Close button - subtle, positioned in header */
-    .interactive-comment-close {
-      position: absolute;
-      top: 10px;
-      right: 10px;
-      width: 24px;
-      height: 24px;
-      border: none;
-      background: transparent;
-      color: var(--pathfinder-comment-close-color, rgba(255, 255, 255, 0.5));
-      border-radius: 4px;
-      cursor: pointer;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 18px;
-      font-weight: normal;
-      line-height: 1;
-      padding: 0;
-      pointer-events: auto;
-      transition: all 0.15s ease;
-      z-index: 1;
-    }
-
-    .interactive-comment-close:hover {
-      background: var(--pathfinder-comment-close-hover-bg, rgba(255, 255, 255, 0.1));
-      color: var(--pathfinder-comment-close-hover-color, rgba(255, 255, 255, 0.9));
-    }
-
-    .interactive-comment-close:active {
-      transform: scale(0.95);
-    }
-
-    /* Legacy skip button - kept for noop steps and backward compatibility */
-    .interactive-comment-skip-btn {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 8px 16px;
-      border: 1px solid var(--pathfinder-comment-btn-border, rgba(255, 255, 255, 0.2));
-      background: transparent;
-      color: var(--pathfinder-comment-btn-text, rgba(255, 255, 255, 0.8));
-      border-radius: 6px;
-      cursor: pointer;
-      font-size: 13px;
-      font-weight: 500;
-      pointer-events: auto;
-      transition: all 0.15s ease;
-      white-space: nowrap;
-    }
-
-    .interactive-comment-skip-btn:hover {
-      background: var(--pathfinder-comment-btn-hover-bg, rgba(255, 255, 255, 0.1));
-      border-color: var(--pathfinder-comment-btn-hover-border, rgba(255, 255, 255, 0.3));
-      color: var(--pathfinder-comment-btn-hover-text, #ffffff);
-    }
-
-    .interactive-comment-skip-btn:active {
-      transform: scale(0.98);
-    }
-
-    /* Button container - navigation row layout */
-    .interactive-comment-buttons {
-      display: flex;
-      align-items: center;
-      gap: 8px;
-      margin-top: 16px;
-      padding-top: 12px;
-      border-top: 1px solid var(--pathfinder-comment-divider, rgba(255, 255, 255, 0.08));
-    }
-
-    /* Form validation status when inline with buttons */
-    .interactive-comment-buttons .interactive-form-validation-status {
-      margin-top: 0;
-      flex: 1;
-      min-width: 120px;
-    }
-
-    /* Legacy cancel button - kept for backward compatibility */
-    .interactive-comment-cancel-btn {
-      display: none; /* Use nav-btn--cancel instead */
-    }
-
-    /* Orange glow border for comment boxes with entrance animation */
-    .interactive-comment-glow {
-      border: 1px solid rgba(255, 136, 0, 0.4) !important;
-      animation: commentGlowEntrance 0.6s ease-out forwards;
-    }
-
-    @keyframes commentGlowEntrance {
-      0% {
-        box-shadow:
-          0 4px 20px rgba(0, 0, 0, 0.25),
-          0 0 0 0 rgba(255, 136, 0, 0),
-          0 0 0 rgba(255, 136, 0, 0);
-      }
-      50% {
-        box-shadow:
-          0 4px 20px rgba(0, 0, 0, 0.25),
-          0 0 0 4px rgba(255, 136, 0, 0.4),
-          0 0 30px rgba(255, 136, 0, 0.3);
-      }
-      100% {
-        box-shadow:
-          0 4px 20px rgba(0, 0, 0, 0.25),
-          0 0 0 2px rgba(255, 136, 0, 0.25),
-          0 0 15px rgba(255, 136, 0, 0.15);
-      }
-    }
-
-    /* Clean content section layout (no logo) */
-    .interactive-comment-content-section {
-      display: flex;
-      flex-direction: column;
-      gap: 8px;
-    }
-
-    /* Legacy wrapper - kept for backward compatibility with simple comments */
-    .interactive-comment-wrapper {
-      display: flex;
-      align-items: flex-start;
-      gap: 8px;
-    }
-
-    .interactive-comment-logo {
-      display: none; /* Hidden by default - clean design without logo */
-    }
-
-    .interactive-comment-text {
-      flex: 1;
-      line-height: 1.4;
-      word-wrap: break-word;
-      overflow-wrap: break-word;
-      max-width: 100%;
-    }
-
-    /* Handle code elements within comments */
-    .interactive-comment-text code,
-    .interactive-comment-description code {
-      display: inline-block;
-      background: var(--pathfinder-comment-code-bg, rgba(255, 255, 255, 0.1));
-      border: 1px solid var(--pathfinder-comment-code-border, rgba(255, 255, 255, 0.2));
-      border-radius: 3px;
-      padding: 1px 4px;
-      font-size: 0.85em;
-      font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-      word-break: break-all;
-      white-space: pre-wrap;
-      max-width: 100%;
-      box-sizing: border-box;
-    }
-
-    /* Handle other inline elements to prevent overflow */
-    .interactive-comment-text strong,
-    .interactive-comment-text em,
-    .interactive-comment-text span,
-    .interactive-comment-description strong,
-    .interactive-comment-description em,
-    .interactive-comment-description span {
-      word-wrap: break-word;
-      overflow-wrap: break-word;
-    }
-
-    /* Arrow using CSS pseudo-element - positioned based on data-position */
-    .interactive-comment-content::before {
-      content: '';
-      position: absolute;
-      width: 0;
-      height: 0;
-    }
-
-    /* Arrow pointing LEFT (comment is to the right of highlight) */
-    .interactive-comment-box[data-position="right"] .interactive-comment-content::before {
-      top: 50%;
-      left: -8px;
-      transform: translateY(-50%);
-      border-top: 8px solid transparent;
-      border-bottom: 8px solid transparent;
-      border-right: 8px solid var(--pathfinder-comment-bg, #1f1f23);
-    }
-
-    /* Arrow pointing RIGHT (comment is to the left of highlight) */
-    .interactive-comment-box[data-position="left"] .interactive-comment-content::before {
-      top: 50%;
-      right: -8px;
-      left: auto;
-      transform: translateY(-50%);
-      border-top: 8px solid transparent;
-      border-bottom: 8px solid transparent;
-      border-left: 8px solid var(--pathfinder-comment-bg, #1f1f23);
-    }
-
-    /* Arrow pointing UP (comment is below highlight) */
-    .interactive-comment-box[data-position="bottom"] .interactive-comment-content::before {
-      top: -8px;
-      left: 50%;
-      transform: translateX(-50%);
-      border-left: 8px solid transparent;
-      border-right: 8px solid transparent;
-      border-bottom: 8px solid var(--pathfinder-comment-bg, #1f1f23);
-    }
-
-    /* Arrow pointing DOWN (comment is above highlight) */
-    .interactive-comment-box[data-position="top"] .interactive-comment-content::before {
-      bottom: -8px;
-      top: auto;
-      left: 50%;
-      transform: translateX(-50%);
-      border-left: 8px solid transparent;
-      border-right: 8px solid transparent;
-      border-top: 8px solid var(--pathfinder-comment-bg, #1f1f23);
-    }
-
-    /* Step checklist in guided comment tooltips (legacy - kept for compatibility) */
-    .interactive-comment-steps-list {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-      margin-bottom: 14px;
-      padding-bottom: 12px;
-      border-bottom: 1px solid var(--pathfinder-comment-border-weak, #404040);
-    }
-
-    .interactive-comment-step-item {
-      font-size: 13px;
-      line-height: 1.4;
-      color: var(--pathfinder-comment-text-secondary, #999999);
-      padding: 2px 0;
-      display: flex;
-      align-items: center;
-      gap: 4px;
-    }
-
-    .interactive-comment-step-item.interactive-comment-step-current {
-      color: var(--pathfinder-comment-text, #d9d9d9);
-      font-weight: 500;
-      background: rgba(255, 193, 7, 0.15);
-      padding: 4px 6px;
-      margin: -2px -4px;
-      border-radius: 3px;
-    }
-
-    /* Comment box header - step badge only */
-    .interactive-comment-header {
-      display: flex;
-      align-items: center;
-      margin-bottom: 12px;
-    }
-
-    .interactive-comment-step-badge {
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      padding: 5px 14px;
-      background: linear-gradient(135deg, #3871dc 0%, #5a67d8 100%);
-      color: #ffffff;
-      border-radius: 20px;
-      font-size: 12px;
-      font-weight: 600;
-      letter-spacing: 0.3px;
-      box-shadow: 0 2px 8px rgba(56, 113, 220, 0.3);
-    }
-
-    /* Progress bar */
-    .interactive-comment-progress-container {
-      position: relative;
-      height: 4px;
-      background: var(--pathfinder-comment-progress-bg, rgba(255, 255, 255, 0.1));
-      border-radius: 4px;
-      overflow: hidden;
-      margin-bottom: 16px;
-    }
-
-    .interactive-comment-progress-bar {
-      position: absolute;
-      top: 0;
-      left: 0;
-      height: 100%;
-      background: linear-gradient(90deg, #3871dc 0%, #5a67d8 100%);
-      border-radius: 4px;
-      transition: width 0.3s ease;
-    }
-
-    /* Step dots - centered, no borders */
-    .interactive-comment-dots {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 6px;
-      margin: 16px 0 8px;
-    }
-
-    .interactive-comment-dot {
-      width: 8px;
-      height: 8px;
-      border-radius: 50%;
-      background: var(--pathfinder-comment-dot-bg, rgba(255, 255, 255, 0.2));
-      transition: all 0.2s ease;
-    }
-
-    .interactive-comment-dot--current {
-      background: #3871dc;
-      box-shadow: 0 0 0 3px rgba(56, 113, 220, 0.3);
-      animation: dotPulse 2s ease-in-out infinite;
-    }
-
-    .interactive-comment-dot--completed {
-      background: #73bf69;
-    }
-
-    @keyframes dotPulse {
-      0%, 100% {
-        box-shadow: 0 0 0 3px rgba(56, 113, 220, 0.3);
-      }
-      50% {
-        box-shadow: 0 0 0 6px rgba(56, 113, 220, 0);
-      }
-    }
-
-    /* Title - clean typography */
-    .interactive-comment-title {
-      font-size: 16px;
-      font-weight: 600;
-      color: var(--pathfinder-comment-text, #ffffff);
-      margin: 0 0 6px 0;
-      line-height: 1.3;
-    }
-
-    /* Description - secondary text */
-    .interactive-comment-description {
-      font-size: 14px;
-      color: var(--pathfinder-comment-text-secondary, rgba(255, 255, 255, 0.75));
-      margin: 0;
-      line-height: 1.5;
-    }
-
-    /* Navigation buttons - clean, minimal */
-    .interactive-comment-nav-btn {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 4px;
-      padding: 8px 16px;
-      border: 1px solid var(--pathfinder-comment-btn-border, rgba(255, 255, 255, 0.2));
-      border-radius: 6px;
-      background: transparent;
-      color: var(--pathfinder-comment-btn-text, rgba(255, 255, 255, 0.8));
-      font-size: 13px;
-      font-weight: 500;
-      cursor: pointer;
-      pointer-events: auto;
-      transition: all 0.15s ease;
-      white-space: nowrap;
-    }
-
-    .interactive-comment-nav-btn:hover:not(:disabled) {
-      background: var(--pathfinder-comment-btn-hover-bg, rgba(255, 255, 255, 0.1));
-      border-color: var(--pathfinder-comment-btn-hover-border, rgba(255, 255, 255, 0.3));
-      color: var(--pathfinder-comment-btn-hover-text, #ffffff);
-    }
-
-    .interactive-comment-nav-btn:disabled {
-      opacity: 0.4;
-      cursor: not-allowed;
-    }
-
-    /* Primary button - gradient */
-    .interactive-comment-nav-btn--primary {
-      background: linear-gradient(135deg, #3871dc 0%, #5a67d8 100%);
-      border-color: transparent;
-      color: #ffffff;
-    }
-
-    .interactive-comment-nav-btn--primary:hover:not(:disabled) {
-      background: linear-gradient(135deg, #4a82e4 0%, #6b74e8 100%);
-      border-color: transparent;
-      transform: translateY(-1px);
-      box-shadow: 0 4px 12px rgba(56, 113, 220, 0.4);
-    }
-
-    /* Cancel button - subtle danger styling */
-    .interactive-comment-nav-btn--cancel {
-      border-color: rgba(255, 100, 100, 0.3);
-      color: rgba(255, 180, 180, 0.9);
-    }
-
-    .interactive-comment-nav-btn--cancel:hover:not(:disabled) {
-      background: rgba(255, 100, 100, 0.15);
-      border-color: rgba(255, 100, 100, 0.5);
-      color: rgba(255, 220, 220, 1);
-    }
-
-    .interactive-comment-nav-spacer {
-      flex: 1;
-    }
-
-    /* Keyboard hint - subtle, integrated */
-    .interactive-comment-keyboard-hint {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 6px;
-      margin-top: 12px;
-      font-size: 11px;
-      color: var(--pathfinder-comment-kbd-text, rgba(255, 255, 255, 0.35));
-    }
-
-    .interactive-comment-kbd {
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      min-width: 20px;
-      padding: 2px 6px;
-      background: var(--pathfinder-comment-kbd-bg, rgba(255, 255, 255, 0.08));
-      border: 1px solid var(--pathfinder-comment-kbd-border, rgba(255, 255, 255, 0.12));
-      border-radius: 4px;
-      font-size: 10px;
-      font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-    }
-
-    /* Hide interactive comment spans - they're extracted as metadata */
-    span.interactive-comment {
-      display: none !important;
-    }
-
-    /* Spinner animation for section running state */
-    @keyframes spin {
-      0% { transform: rotate(0deg); }
-      100% { transform: rotate(360deg); }
-    }
-
-    /* Form validation status in comment box (for guided formfill steps) */
-    .interactive-form-validation-status {
-      margin-top: 10px;
-      padding: 8px 10px;
-      border-radius: 4px;
-      font-size: 12px;
-      display: flex;
-      align-items: center;
-      gap: 6px;
-    }
-
-    .interactive-form-validation-status.form-checking {
-      background: var(--pathfinder-form-checking-bg, rgba(255, 255, 255, 0.1));
-      border: 1px solid var(--pathfinder-form-checking-border, rgba(255, 255, 255, 0.2));
-      color: var(--pathfinder-form-checking-text, rgba(255, 255, 255, 0.8));
-    }
-
-    .interactive-form-validation-status.form-hint-warning {
-      background: rgba(255, 193, 7, 0.2);
-      border: 1px solid rgba(255, 193, 7, 0.4);
-      color: rgba(255, 220, 150, 1);
-    }
-
-    .interactive-form-validation-status.form-valid {
-      background: rgba(115, 191, 105, 0.2);
-      border: 1px solid rgba(115, 191, 105, 0.4);
-      color: rgba(115, 191, 105, 1);
-    }
-
-    .interactive-form-validation-status .interactive-form-spinner {
-      animation: spin 1s linear infinite;
-    }
-
-    .interactive-form-validation-status .interactive-form-warning-icon {
-      flex-shrink: 0;
-    }
-
-    .interactive-form-validation-status .interactive-form-success-icon {
-      flex-shrink: 0;
-      animation: popIn 0.3s ease-out;
-    }
-
-    @keyframes popIn {
-      0% {
-        transform: scale(0);
-        opacity: 0;
-      }
-      50% {
-        transform: scale(1.2);
-      }
-      100% {
-        transform: scale(1);
-        opacity: 1;
-      }
-    }
-  `;
-  document.head.appendChild(style);
-};
